+++ package.json
@@ -25,6 +25,7 @@
     "postcss": "^8.4.0",
     "svelte": "^5.0.0",
     "svelte-check": "^4.0.0",
+    "svelte-eslint-parser": "^1.6.0",
     "tailwindcss": "^3.4.0",
     "typescript": "^5.0.0",
     "vite": "^6.0.0"
